Update for v2 version of ably-js
Changes how we import the library in worker.js.
Adds an example on how to use a modular variant of the library from v2.

I couldn't manage to change `ably` dependency in package.json to
reference an 'integrations/v2' branch of `ably-js` repo on github -
it would fail due to git submodules (we are using ably-common in there).
So for now to test with v2 version of the library we should refer to
"Testing a non-published version of `ably-js`" in README.